:art: refactor(View : MoveResultDisplay) : 출력 관련 enum 데이터로 변환하는 기능을 별 상관없는 Util 클래스가 아닌 enum 자신이 가지도록 개선
